requirements.txt
numpy
boto3
botocore
Pillow
pyyaml
